fix: SignalHandler not exported in mobile package
